エラー攻撃関係のスクリプトの修正
