--- Big Wave prototype/Assets/Script/EnemyBulletScript/WallBullet.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 519bdbe8564ae6343955a837743ab857
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
