+++ Big Wave prototype/Assets/Script/EnemyBulletScript/WallBulletScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed8b5f584ecc1134e8b419a0f6742cc9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
