+++ Big Wave prototype/Assets/Script/EnemyScript/EnemyActionType/EnemyActionTypeWallScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60b9aa05782a4294489273d8897acfcb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
