+++ Big Wave prototype/Assets/Script/EnemyScript/EnemyActionType/EnemyActionTypeWallScripts/EnemyActionTypeShotWall.cs
@@ -1,0 +1,75 @@
+using Microsoft.Win32.SafeHandles;
+using System.Collections;
+using System.Collections.Generic;
+using Unity.VisualScripting;
+using UnityEngine;
+
+//作成者：桑原
+
+public partial class EnemyActionTypeShotWall : EnemyActionTypeBase
+{
+    public override void OnEnter(EnemyActionTypeBase[] beforeActionType)
+    {
+        Debug.Log("Wall");
+
+        currentDelayTime = 0;
+
+        actionEffect.Generate();//エフェクト生成
+
+        animController.AnimControl_Trigger(animName);
+    }
+
+    public override void OnUpdate()
+    {
+        if (!shoted)
+        {
+            if (wallAreaInstance == null)
+            {
+                GenerateWallArea();
+            }
+
+            else
+            {
+                Shot();
+            }
+        }
+    }
+
+    public override void OnExit(EnemyActionTypeBase[] nextActionType)
+    {
+        shoted = false;
+    }
+
+    void GenerateWallArea()
+    {
+        if (wallAreaInstance == null)
+        {
+            //攻撃を撃ちだす位置を取得
+            Vector3 shotPos = shotPosObject.transform.position;
+            Quaternion shotRot = shotPosObject.transform.rotation;
+
+            wallAreaInstance = Instantiate(wallAreaPrefab, shotPos, shotRot, gamePos.transform);
+
+            WallBullet wallBulletScript = wallAreaInstance.AddComponent<WallBullet>();
+
+            //ToggleColliderOfWallBulletにwallBulletの参照を設定
+            wallBulletScript.SetWallBullet(this);
+
+            //弾のRigidbodyを取得
+            bulletRb = wallAreaInstance.GetComponent<Rigidbody>();
+        }
+    }
+
+    void Shot() //弾を撃つ
+    {
+        currentDelayTime += Time.deltaTime;
+
+        if (currentDelayTime > delayTime)
+        {
+            //弾を撃ちだす
+            bulletRb.AddForce(-transform.forward * shotPower, ForceMode.Impulse);
+
+            shoted = true;
+        }
+    }
+}
