+++ Big Wave prototype/Assets/Script/EnemyScript/EnemyActionType/EnemyActionTypeWallScripts/ShotWallParameters.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 7156604cd024931478d7886b2e86d310
+guid: 8dd83beb1b7118d47a4e21cbb833b7a7
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
